Added ResourceReference.makeDirectory and makeFile methods; use in elFinder connector for corresponding commans

+++ framework/src/main/groovy/org/moqui/impl/context/reference/UrlResourceReference.groovy
@@ -122,7 +122,9 @@ class UrlResourceReference extends BaseResourceReference {
         }
     }
 
+    @Override
     boolean supportsLastModified() { isFileProtocol }
+    @Override
     long getLastModified() {
         if (isFileProtocol) {
             return getFile().lastModified()
@@ -131,11 +133,16 @@ class UrlResourceReference extends BaseResourceReference {
         }
     }
 
+    @Override
     boolean supportsSize() { isFileProtocol }
+    @Override
     long getSize() { isFileProtocol ? getFile().length() : 0 }
 
+    @Override
     boolean supportsWrite() { isFileProtocol }
+    @Override
     void putText(String text) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not support for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
         // first make sure the directory exists that this is in
         if (!getFile().parentFile.exists()) getFile().parentFile.mkdirs()
         // now write the text to the file and close it
@@ -144,7 +151,9 @@ class UrlResourceReference extends BaseResourceReference {
         fw.close()
         this.exists = null
     }
+    @Override
     void putStream(InputStream stream) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not support for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
         // first make sure the directory exists that this is in
         if (!getFile().parentFile.exists()) getFile().parentFile.mkdirs()
         OutputStream os = new FileOutputStream(getFile())
@@ -154,6 +163,7 @@ class UrlResourceReference extends BaseResourceReference {
         this.exists = null
     }
 
+    @Override
     void move(String newLocation) {
         if (!newLocation) throw new IllegalArgumentException("No location specified, not moving resource at ${getLocation()}")
         ResourceReference newRr = ecf.resource.getLocationReference(newLocation)
@@ -165,5 +175,22 @@ class UrlResourceReference extends BaseResourceReference {
 
         String path = newRr.getUrl().toExternalForm().substring(5)
         getFile().renameTo(new File(path))
+    }
+
+    @Override
+    ResourceReference makeDirectory(String name) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not support for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
+        UrlResourceReference newRef = (UrlResourceReference) new UrlResourceReference().init("${location}/${name}", ecf)
+        newRef.getFile().mkdirs()
+        return newRef
+    }
+    @Override
+    ResourceReference makeFile(String name) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not support for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
+        UrlResourceReference newRef = (UrlResourceReference) new UrlResourceReference().init("${location}/${name}", ecf)
+        // first make sure the directory exists that this is in
+        if (!getFile().exists()) getFile().mkdirs()
+        newRef.getFile().createNewFile()
+        return newRef
     }
 }
